pyproject.toml
[tool.poetry]
name = "formation"
version = "0.1.27"
authors = ["Dotan Nahum <jondotan@gmail.com>"]
description = "A generic functional middleware infrastructure for Python."

license = "MIT"

readme = "README.md"

homepage = "https://github.com/jondot/formation"
repository = "https://github.com/jondot/formation"
documentation = "https://github.com/jondot/formation"

keywords = ["flask", "django", "requests", "middleware", "functional"]

[tool.poetry.dependencies]
python = "*"
toolz = "^0.9.0"
cytoolz = "^0.9.0"
requests = "^2.20"
attrs = "^18.2"
pybreaker = "^0.4.5"
xmltodict = "^0.11.0"
lxml = "^4.2"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.6"
pytest-sugar = "^0.9.2"
boltons = "^18.0"
pytest = "^3.0"
pytest-benchmark = "^3.1"
pytest-testmon = "^0.9.13"
pytest-watch = "^4.2"
snapshottest = "^0.5.0"

flake8 = { version = "^3.5" }
requests = "^2.20"
pytest-vcr = "^1.0"
attrs-serde = "^0.2.1"
attrs = "^18.2"
structlog = "^18.2"

[tool.poetry.extras]
lint = ["flake8-mypy"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"